jrpc: correctly handle uints
Earlier versions of this didn't handle uint parsing, at all.
As a result, structs that would be corretly parsed by the json.Unmarshal
package were failing to be parsed by the 'custom' param parser. Which
was pretty bad, since it meant we can't reuse structs between
the lightning RPC and the plugin events if they containted uint's.

This fixes that issue, simply, by extending the logic we used for
ints.