jrpc2: handle string->[]byte & RawMessages
Allows us to properly convert strings of raw bytes into byte fields
and provides an escape hatch for 'RawMessage' types to be reconverted
into JSON and then parsed.

The RawMessages hack is pretty inefficient because we parse it into type
less objects, then back to JSON, then eventually into the object that we
want.

This is only used by the rpc_command hook, as it needs to be able to
figure out what command to parse it into before attempting to parse it.

FIXME: maybe take more advantage of the RawMessage parser?